+++ app/src/main/java/com/maalelan/postcardstorehouse/fragments/AddPostcardFragment.java
@@ -1,0 +1,10 @@
+package com.maalelan.postcardstorehouse.fragments;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+public class AddPostcardFragment extends Fragment {
+
+
+}